Extend JsonReader and VoldemortClientShell to handle integer types up to int64 (long) inste\ ad of just int32 (int).
To test directly in the shell, I changed the stores.xml to use a java-serialization value-s\
erializer:
-      <type>string</type>
+      <type>java-serialization</type>
I am not committing this change to stores.xml though.

I updated the JsonReaderTest to make all int constant values longs so the test passes.

+++ src/java/voldemort/serialization/json/JsonReader.java
@@ -248,7 +248,7 @@ public class JsonReader {
 
     public Number readNumber() {
         skipWhitespace();
-        int intPiece = readInt();
+        long intPiece = readLong();
 
         // if int is all we have, return it
         if(isTerminator(current()))
@@ -261,7 +261,7 @@ public class JsonReader {
         if(current() == 'e' || current() == 'E') {
             next();
             skipIf('+');
-            int frac = readInt();
+            long frac = readLong();
             doublePiece *= Math.pow(10, frac);
         }
         if(isTerminator(current()))
@@ -276,9 +276,9 @@ public class JsonReader {
                || ch == ',' || ch == -1;
     }
 
-    public int readInt() {
+    public long readLong() {
         skipWhitespace();
-        int val = 0;
+        long val = 0;
         boolean isPositive;
         if(current() == '-') {
             isPositive = false;
